num4
cambios en las relaciones del modelo

+++ src/main/java/com/Caso1Backend/back/security/repository/FacturaCabeceraRepository.java
@@ -4,6 +4,6 @@ import com.Caso1Backend.back.security.models.FacturaCabecera;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaCabeceraRepository extends JpaRepository<FacturaCabecera, Long> {
+public interface FacturaCabeceraRepository extends JpaRepository<FacturaCabecera, Integer> {
     
 }
