+++ src/main/java/com/Caso1Backend/back/security/repository/FacturaCuerpoRepository.java
@@ -4,6 +4,6 @@ import com.Caso1Backend.back.security.models.FacturaCuerpo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaCuerpoRepository extends JpaRepository<FacturaCuerpo, Long> {
+public interface FacturaCuerpoRepository extends JpaRepository<FacturaCuerpo, Integer> {
     
 }
